Move mtga_follower.py directly under Python source directory
